+++ api/server.go
@@ -90,6 +90,8 @@ type GRPCServer struct {
 	marketDataStore    *sqlstore.MarketData
 	tradeStore         *sqlstore.Trades
 	assetStore         *sqlstore.Assets
+	accountStore       *sqlstore.Accounts
+	marketsStore       *sqlstore.Markets
 
 	eventObserver *eventObserver
 
@@ -136,6 +138,8 @@ func NewGRPCServer(
 	marketDataStore *sqlstore.MarketData,
 	tradeStore *sqlstore.Trades,
 	assertStore *sqlstore.Assets,
+	accountStore *sqlstore.Accounts,
+	marketsStore *sqlstore.Markets,
 ) *GRPCServer {
 	// setup logger
 	log = log.Named(namedLogger)
@@ -179,6 +183,8 @@ func NewGRPCServer(
 		marketDataStore:         marketDataStore,
 		tradeStore:              tradeStore,
 		assetStore:              assertStore,
+		accountStore:            accountStore,
+		marketsStore:            marketsStore,
 		eventObserver: &eventObserver{
 			log:          log,
 			eventService: eventService,
@@ -327,6 +333,9 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 			orderStore:         g.orderStore,
 			tradeStore:         g.tradeStore,
 			assetStore:         g.assetStore,
+			accountStore:       g.accountStore,
+			marketDataStore:    g.marketDataStore,
+			marketsStore:       g.marketsStore,
 		}
 	} else {
 		g.tradingDataService = tradingDataSvc