+++ api/trading_test.go
@@ -277,6 +277,8 @@ func getTestGRPCServer(
 	sqlMarketDataStore := sqlstore.NewMarketData(&sqlStore)
 	sqlTradeStore := sqlstore.NewTrades(&sqlStore)
 	sqlAssetStore := sqlstore.NewAssets(&sqlStore)
+	sqlAccountStore := sqlstore.NewAccounts(&sqlStore)
+	sqlMarketsStore := sqlstore.NewMarkets(&sqlStore)
 
 	g := api.NewGRPCServer(
 		logger,
@@ -315,6 +317,8 @@ func getTestGRPCServer(
 		sqlMarketDataStore,
 		sqlTradeStore,
 		sqlAssetStore,
+		sqlAccountStore,
+		sqlMarketsStore,
 	)
 	if g == nil {
 		err = fmt.Errorf("failed to create gRPC server")