+++ sqlstore/market_data.go
@@ -8,6 +8,7 @@ import (
 	"time"
 
 	"code.vegaprotocol.io/data-node/entities"
+	"code.vegaprotocol.io/data-node/logging"
 	"github.com/georgysavva/scany/pgxscan"
 )
 
@@ -16,7 +17,7 @@ type MarketData struct {
 }
 
 const (
-	sqlColumns = `market, vega_time, seq_num, mark_price, 
+	sqlMarketDataColumns = `market, vega_time, seq_num, mark_price, 
 		best_bid_price, best_bid_volume, best_offer_price, best_offer_volume,
 		best_static_bid_price, best_static_bid_volume, best_static_offer_price, best_static_offer_volume,
 		mid_price, static_mid_price, open_interest, auction_end, 
@@ -44,7 +45,7 @@ func (md *MarketData) Add(data *entities.MarketData) error {
 			$13, $14, $15, $16, 
 			$17, $18, $19, $20, 
 			$21, $22, $23, $24, 
-			$25, $26, $27)`, sqlColumns)
+			$25, $26, $27)`, sqlMarketDataColumns)
 
 	if _, err := md.pool.Exec(ctx, query,
 		data.Market, data.VegaTime, data.SeqNum, data.MarkPrice,
@@ -62,16 +63,28 @@ func (md *MarketData) Add(data *entities.MarketData) error {
 	return nil
 }
 
-func (md *MarketData) GetByID(ctx context.Context, marketID string) (entities.MarketData, error) {
+func (md *MarketData) GetMarketDataByID(ctx context.Context, marketID string) (entities.MarketData, error) {
+	md.log.Debug("Retrieving market data from Postgres", logging.String("market-id", marketID))
 	market, err := hex.DecodeString(marketID)
 	if err != nil {
 		return entities.MarketData{}, fmt.Errorf("bad ID (must be a hex string): %w", err)
 	}
 
 	var marketData entities.MarketData
-	query := fmt.Sprintf("select %s from market_data_snapshot where market = $1", sqlColumns)
+	query := fmt.Sprintf("select %s from market_data_snapshot where market = $1", sqlMarketDataColumns)
 
 	err = pgxscan.Get(ctx, md.pool, &marketData, query, market)
+
+	return marketData, err
+}
+
+func (md *MarketData) GetMarketsData(ctx context.Context) ([]entities.MarketData, error) {
+	md.log.Debug("Retrieving markets data from Postgres")
+
+	var marketData []entities.MarketData
+	query := fmt.Sprintf("select %s from market_data_snapshot", sqlMarketDataColumns)
+
+	err := pgxscan.Select(ctx, md.pool, &marketData, query)
 
 	return marketData, err
 }
@@ -100,7 +113,7 @@ func (md *MarketData) getBetweenDatesByID(ctx context.Context, marketID string, 
 		return nil, err
 	}
 
-	selectStatement := fmt.Sprintf(`select %s from market_data`, sqlColumns)
+	selectStatement := fmt.Sprintf(`select %s from market_data`, sqlMarketDataColumns)
 
 	if start != nil && end != nil {
 		query, args := orderAndPaginateQuery(