+++ CHANGELOG.md
@@ -15,6 +15,18 @@
 - [](https://github.com/vegaprotocol/data-node/pull/xxx) -
 
 
+## 0.49.2
+
+### 🛠  Improvements
+- [404](https://github.com/vegaprotocol/data-node/pull/404) - Migrate market data API to retrieve data from `Postgres`
+- [406](https://github.com/vegaprotocol/data-node/pull/406) - Add a basic integration test
+- [412](https://github.com/vegaprotocol/data-node/pull/412) - Migrate markets API to retrieve data from `Postgres`
+- [407](https://github.com/vegaprotocol/data-node/pull/407) - Add `positionDecimalPlaces` to market `graphQL`
+
+### 🐛 Fixes
+- [411](https://github.com/vegaprotocol/data-node/pull/411) - Fix a couple of incompatibilities in `data-node v2`
+- [417](https://github.com/vegaprotocol/data-node/pull/411) - Report correct total tokens for a vote in `graphql`
+
 ## 0.49.1
 
 ### 🚨 Breaking changes
@@ -24,6 +36,7 @@
 - [362](https://github.com/vegaprotocol/data-node/pull/362) - Added support using TLS for `GraphQL` connections
 - [393](https://github.com/vegaprotocol/data-node/pull/393) - Data store migration
 - [395](https://github.com/vegaprotocol/data-node/pull/395) - Migrate Asset API to retrieve data from `Postgres`
+- [399](https://github.com/vegaprotocol/data-node/pull/399) - Migrate Accounts API to retrieve data from `Postgres`
 
 ### 🐛 Fixes
 - [387](https://github.com/vegaprotocol/data-node/pull/387) - Fixes incorrect data types in the `MarketData` proto message